==> 1.3. ETL items.ipynb <==
import gzip
import json
import csv
import pandas as pd
import ast

from pandas import json_normalize

import my_functions
%load_ext autoreload
%autoreload 2

Se realiza la carga del archivo 'australian_users_items.json' en el Dataframe 'df_items'.

with open("Datasets/Datasets Originales/australian_users_items.json", 'r', encoding='utf-8') as data:
    data = data.readlines()
    
data = [eval(line) for line in data]

df_items = pd.DataFrame(data)

df_items.head(3)

Se guarda una versión en crudo en el archivo 'user_items_raw.csv'

df_items.to_csv('Datasets/Datasets ETL/user_items_raw.csv', index=False, encoding='utf-8')

df_items = pd.read_csv('Datasets/Datasets ETL/user_items_raw.csv')

Se comienza la exploración del dataframe.

df_items.info()

Campo 'items'

En este campo se encuentran una lista con items (juegos) que posee cada usuario, cada item contenido en un diccionario junto a otra información.

Se procederá a desanidar este campo de forma tal que contaremos con una fila por cada item que posea un usuario.

# Desanidar la columna 'items'
df_items = pd.json_normalize(data, record_path=['items'], meta=['steam_id','items_count','user_id', 'user_url'] )
df_items.head(3)

Habiendo desanidado la columna 'items' se generaron tantas filas para cada 'user_id' como items posea. Siendo que existen usuarios que poseen muchos juegos, los registros se multiplicaron considerablemente.

df_items.info()

Se explora en busca de duplicados.

# Contar filas duplicadas
total_duplicadas = df_items.duplicated().sum()

# Mostrar el total de filas duplicadas
print(f"Total de filas duplicadas: {total_duplicadas}")

Se eliminan más de 59 mil registros duplicados.

df_items = df_items.drop_duplicates(keep='first')

my_functions.describe_df(df_items)

Tras una primera exploración se determina que los campos 'items_count', 'user_url' y 'playtime_2weeks' no son necesarios, por lo que se los descarta.

# Lista de columnas a eliminar
columnas_a_eliminar = ['items_count', 'user_url', 'playtime_2weeks']

# Realizar el drop de las columnas
df_items = df_items.drop(columnas_a_eliminar, axis=1)


# Mostrar el DataFrame actualizado
df_items.head(3)

Como se realizó en los archivos anteriores, se modifica el nombre de la columna 'item_id' por 'game_id', siendo este el campo compartido por los 3 datasets originales con los que se cuenta.

También se modifica el nombre del campo 'item_name' por 'app_name', correspondiéndose el mismo con el campo 'app_name' de el dataset trabajado en '1.1 ETL games'.

# Cambiar el nombre de la columna 'item_id' a 'game_id'
df_items.rename(columns={'item_id': 'game_id'}, inplace=True)

# Cambiar el nombre de la columna 
df_items.rename(columns={'item_name': 'app_name'}, inplace=True)

Se realiza una comprobación de registros duplicados y al hallar algunos se procede a eliminarlos. Al observar que la única diferencia entre los registros es que algunos tienen un 'playtime_forever' mayor (para el mismo usuario y mismo juego) se eliminan los primeros registros, quedándose aquellos con mayor tiempo de juego.

# Filtrar filas duplicadas
duplicados = df_items[df_items.duplicated(subset=['game_id', 'user_id'], keep=False)]

# Ordenar por game_id y user_id
duplicados = duplicados.sort_values(by=['game_id', 'user_id'])

# Cantidad de filas duplicadas
print('Total filas duplicadas: ', len(duplicados))

# Mostrar resultado
duplicados


# Eliminar duplicados y quedarse con el último registro de cada duplicado
df_items = df_items.drop_duplicates(subset=['game_id', 'user_id'], keep='last')

Se verifican modificaciones.

# Filtrar filas duplicadas
duplicados = df_items[df_items.duplicated(subset=['game_id', 'user_id'], keep=False)]

# Ordenar por game_id y user_id
duplicados = duplicados.sort_values(by=['game_id', 'user_id'])

# Cantidad de filas duplicadas
print('Total filas duplicadas: ', len(duplicados))

# Mostrar resultado
duplicados

Se determina eliminar el campo 'steam_id', considerando que el 'user_id' será más útil dado que es un campo compartido con el dataset de reseñas.

# Eliminar 'steam_id'
df_items = df_items.drop('steam_id', axis=1)

Se explora una vez más la estructura de los datos. 
Se determina que los mismos se encuentran en condiciones de ser utilizados para los objetivos propuestos y se guardan en un nuevo archivo, 'user_items_v1.csv'.

my_functions.describe_df(df_items)


Se guarda el dataset con los datos limpios.

df_items.to_csv('Datasets/Datasets ETL/user_items_v1.csv', index=False, encoding='utf-8')
