==> 3. ETL merge.ipynb <==
import csv
import pandas as pd
import numpy as np
import pyarrow as pa
import pyarrow.parquet as pq

import my_functions
%load_ext autoreload
%autoreload 2

## Carga de datos

En esta etapa del proyecto se tomaran los 3 dataset ya trabajados en el ETL, utilizando la información necesaria para conformar otros dataset que contengan únicamente los datos necesarios para procesar cada una de las consultas solicitadas para al API.

Se realiza la carga de los dataset df_games, df_reviews y df_items en dataframes.

# Cargar el archivo CSV en un DataFrame
df_games = pd.read_csv('Datasets/Datasets ETL/steam_games_v2.csv')

# Cargar el archivo CSV en un DataFrame
df_reviews = pd.read_csv('Datasets/Datasets ETL/user_reviews_v1.csv')

# Cargar el archivo CSV en un DataFrame
df_items = pd.read_csv('Datasets/Datasets ETL/user_items_v2.csv')

Se examinan brevemente para verificar la carga.

# Mostrar primera fila del DataFrame
df_games.head(1)

my_functions.describe_df(df_games)

# Modificar tipo de columna 'id' a string
df_games['game_id'] = df_games['game_id'].astype(str)

# Mostrar la primera fila del DataFrame
df_reviews.head(1)

my_functions.describe_df(df_reviews)

# Mostra la primera fila del DataFrame
df_items.head(1)

my_functions.describe_df(df_items)

## Creacion de DataFrames auxiliares

En este punto se comenzará con el proceso para obtener los dataset finales, para lo cual se crearán algunos dataframe auxiliares.

## df_play_gnr_top

Este dataframe contendrá los campos 'genre', 'release_year'	y 'playtime' agrupados. Es decir, contiene los valores totales de tiempo de juego ('playtime'), para cada género en un año dado.

El mismo será utilizado en la función que se creará más adelante 'def PlayTimeGenre', que responderá a consultas sobre el año con más horas jugadas para cada género.

Para ello se utilizan campos de df_games y df_items, generando el datafram con la información agrupada como se requiere para el objetivo.

# Modificar tipo de columna 'id' a string
df_games['game_id'] = df_games['game_id'].astype(int)

# Realizar la fusión (merge) entre df_games y df_items en base al campo game_id
merged_df = pd.merge(df_games, df_items, on='game_id', how='inner')

# Inicializar un diccionario para almacenar los datos
playtime_data = {'genre': [], 'release_year': [], 'playtime': []}

# Obtener la lista de géneros (columnas que comienzan con prefijo 'gnr_')
genres = [col for col in df_games.columns if col.startswith('gnr_')]

# Iterar sobre los géneros
for genre in genres:
    # Filtrar df_games para obtener los game_id con el género actual
    genre_game_ids = df_games[df_games[genre] == 1]['game_id']

    # Filtrar el DataFrame fusionado por los game_id con el género actual
    genre_data = merged_df[merged_df['game_id'].isin(genre_game_ids)]

    # Agrupar por género y año, sumar el playtime_forever
    grouped_data = genre_data.groupby(['release_year']).agg({'playtime_forever': 'sum'}).reset_index()

    # Agregar datos al diccionario
    for index, row in grouped_data.iterrows():
        # Quitar el prefijo 'gnr_' de la columna genre
        genre_name = genre.replace('gnr_', '')
        playtime_data['genre'].append(genre_name)
        playtime_data['release_year'].append(row['release_year'])
        playtime_data['playtime'].append(row['playtime_forever'])

# Crear el DataFrame df_play_gnr
df_play_gnr = pd.DataFrame(playtime_data)

# Ordenar por género alfabéticamente y playtime de manera descendente
df_play_gnr = df_play_gnr.sort_values(by=['genre', 'playtime'], ascending=[True, False])

# Restablecer el índice después de la clasificación
df_play_gnr = df_play_gnr.reset_index(drop=True)

Se verifica la carga de los datos visualizando los primeros 3 resultados para cada género, que ya fueron ordenados por el campo 'playtime' de mayor a menor.

pd.set_option('display.max_rows', None)

# Imprimir los primeros 3 resultados para cada género
df_play_gnr.groupby('genre').head(3)

Por último, se genera un nuevo dataframe con la información exclusivamente necesaria para responder a la consulta y se guarda en el archivo 'df_play_gnr_top.csv'.

# Crear nuevo dataframe primer resultado para cada género
df_play_gnr_top = df_play_gnr.groupby('genre').head(1).reset_index(drop=True)

# Mostrar dataframe
pd.reset_option('display.max_rows')
df_play_gnr_top

# Exportar a formato XLSX
# df_play_gnr_top.to_excel('Datasets/Datasets final/df_play_gnr.xlsx', index=False)

# Exportar a formato CSV
df_play_gnr_top.to_csv('Datasets/Datasets final/df_play_gnr.csv', index=False)

# Exportar a formato Parquet
# df_play_gnr_top.to_parquet('Datasets/Datasets final/df_play_gnr.parquet', index=False)

## df_usr_gnr_top / df_usr_gnr_top_year

Se crearán 2 dataframes referidos a tiempo de juego de los usuarios por género.
El primero ('df_usr_gnr_top'), con datos de el usuario con más tiempo de juego para cada género.
El segundo ('df_usr_gnr_top_year'), para cada usuario referido en el dataframe anterior, contendrá los tiempos de juego por año para dicho género.

Primero se conforma un dataframe auxiliar con información de 'df_games' y 'df_items'.

# Hacer un merge entre df_items y df_games y seleccionar campos específicos
df_usr_gnr = pd.merge(df_items, df_games[['game_id', 'release_year'] + [col for col in df_games.columns if col.startswith('gnr_')]], on='game_id', how='inner')[['user_id', 'game_id', 'playtime_forever', 'release_year'] + [col for col in df_games.columns if col.startswith('gnr_')]]

# Mostrar df_usr_items_gnr
df_usr_gnr.head()

Se procede a generar registros diferenciados por género para cada juego. Si un juego contiene más de un género, se multiplicarán los registros.

# Utilizar melt para desplegar las columnas de género
df_usr_gnr = pd.melt(df_usr_gnr, id_vars=['user_id', 'game_id', 'release_year', 'playtime_forever'], var_name='genre', value_name='genre_value')

# Filtrar las filas con valor 1 en la columna de género
df_usr_gnr = df_usr_gnr[df_usr_gnr['genre_value'] == 1]

# Filtrar las filas con valor 0 en la columna playtime
df_usr_gnr = df_usr_gnr[df_usr_gnr['playtime_forever'] != 0]

# Eliminar la columna de valor de género
df_usr_gnr = df_usr_gnr.drop(columns=['genre_value'])

# Quitar el prefijo 'gnr_' de la columna genre
df_usr_gnr['genre'] = df_usr_gnr['genre'].str.replace('gnr_', '')

# Agrupar por user_id, release_year y genre, sumando playtime_forever
df_usr_gnr = df_usr_gnr.groupby(['user_id', 'release_year', 'genre']).agg({'playtime_forever': 'sum'}).reset_index()

# Ordenar el DataFrame
df_usr_gnr = df_usr_gnr.sort_values(by=['genre', 'playtime_forever'], ascending=[True, False]).reset_index(drop=True)

# Mostrar df_usr_gnr
df_usr_gnr.head(5)

Se verifica la estrucutra del dataframe.

df_usr_gnr.info()

df_usr_gnr['genre'].unique()

# Exportar a formato XLSX
# df_usr_gnr.to_excel('Datasets/df_usr_gnr.xlsx', index=False)

# Exportar a formato CSV
# df_usr_gnr.to_csv('Datasets/Datasets merge/df_usr_gnr.csv', index=False)

# Exportar a formato Parquet
# df_usr_gnr.to_parquet('Datasets/Datasets merge/df_usr_gnr.parquet', index=False)

A continuación se agrupan los datos por usuario y género para obtener posteriormente el usuario con más horas jugada para cada género.

# Agrupar por user_id, y genre, sumando playtime_forever
df_usr_gnr_total = df_usr_gnr.groupby(['user_id', 'genre']).agg({'playtime_forever': 'sum'}).reset_index().sort_values(by=['genre', 'playtime_forever'], ascending=[True, False])

# Mostrar df_usr_gnr
df_usr_gnr_total

# Exportar a formato XLSX
# df_usr_gnr_total.to_excel('Datasets/Datasets merge/df_usr_gnr_total.xlsx', index=False)

# Exportar a formato CSV
# df_usr_gnr_total.to_csv('Datasets/Datasets merge/df_usr_gnr_total.csv', index=False)

# Exportar a formato Parquet
# df_usr_gnr_total.to_parquet('Datasets/Datasets merge/df_usr_gnr_total.parquet', index=False)

Utilizando este último dataframe se extraen los datos necesario para responder la consulta y se guardan en 'df_usr_gnr_top.csv'.

# Copiar df_usr_gnr_total a df_usr_gnr_top
df_usr_gnr_top = df_usr_gnr_total.groupby('genre').first().reset_index()

# Mostrar df_usr_gnr_top
df_usr_gnr_top


# Exportar a formato XLSX
# df_usr_gnr_top.to_excel('Datasets/Datasets final/df_usr_gnr_top.xlsx', index=False)

# Exportar a formato CSV
df_usr_gnr_top.to_csv('Datasets/Datasets final/df_usr_gnr_top.csv', index=False)

# Exportar a formato Parquet
# df_usr_gnr_top.to_parquet('Datasets/Datasets final/df_usr_gnr_top.parquet', index=False)

Por último, se genera 'df_usr_gnr_top_year.csv' con la información de tiempo de juego por año de los usuarios con más horas de juego para cada género.

Este dataset contendrá información de tiempo de juego por año ('playtime_forever') y el tiempo de juego acumulado a lo largo de los años ('playtime_acumulated').

# Copiar df_usr_gnr_top a df_usr_gnr_top_year
df_usr_gnr_top_year = pd.merge(df_usr_gnr_top[['user_id', 'genre']], df_usr_gnr[['user_id', 'genre', 'release_year', 'playtime_forever']], on=['user_id', 'genre'], how='left')

# Ordenar el DataFrame
df_usr_gnr_top_year = df_usr_gnr_top_year.sort_values(by=['genre', 'release_year'], ascending=[True, True])

# Crear una nueva columna playtime_acumulated que sume el acumulado por género y usuario
df_usr_gnr_top_year['playtime_acumulated'] = df_usr_gnr_top_year.groupby(['user_id', 'genre'])['playtime_forever'].cumsum()

df_usr_gnr_top_year = df_usr_gnr_top_year.reset_index()

# Seleccionar las columnas requeridas (user_id, genre y playtime_acumulated)
# df_usr_gnr_top_year = df_usr_gnr_top_year[['user_id', 'genre', 'playtime_acumulated']]

# Mostrar df_usr_gnr_top_year
pd.set_option('display.max_rows', None)
df_usr_gnr_top_year


# Exportar a formato XLSX
# df_usr_gnr_top_year.to_excel('Datasets/Datasets final/df_usr_gnr_top_year.xlsx', index=False)

# Exportar a formato CSV
df_usr_gnr_top_year.to_csv('Datasets/Datasets final/df_usr_gnr_top_year.csv', index=False)

# Exportar a formato Parquet
# df_usr_gnr_top_year.to_parquet('Datasets/Datasets final/df_usr_gnr_top_year.parquet', index=False)

## df_games_rec

Este dataframe auxiliar servirá de base para procesar la información necesaria de la consulta 'def UsersRecommend', que debe devolver el top 3 de juegos más recomendados para el año dado.
Se define que un juego es recomendado por un usuario a partir de los campos 'recommend' (si es 'True') y 'sentiment' (positivos / neutrales) de el dataframe df_reviews.


# Crear df_user_rec con las columnas requeridas
df_games_rec = pd.merge(df_reviews[df_reviews['recommend'] & df_reviews['sentiment'].isin([1, 2])],
                        df_games[['game_id']],
                        on='game_id',
                        how='left')

# Agrupar por game_id y posted_year, calcular total_recommend
df_games_rec = df_games_rec.groupby(['game_id', 'posted_year']).agg(total_recommend=('recommend', 'sum')).reset_index()

# Ordenar el DataFrame
df_games_rec = df_games_rec.sort_values(by=['posted_year', 'total_recommend'], ascending=[True, False])

# Agregar app_name al df_games_rec tomando los app_name de df_games
df_games_rec = pd.merge(df_games_rec, df_games[['game_id', 'app_name']],
                        on='game_id', how='left', suffixes=('', '_games')).reset_index()

# Manejar el caso de game_id en df_reviews que no está en df_games
df_games_rec['app_name'].fillna('Sin datos', inplace=True)

# Mostrar primeros 3 registros por año
df_games_rec.groupby('posted_year').head(3)

# Exportar a formato XLSX
# df_games_rec.to_excel('Datasets/Datasets merge/df_games_rec.xlsx', index=False)

# Exportar a formato CSV
# df_games_rec.to_csv('Datasets/Datasets merge/df_games_rec.csv', index=False)

# Exportar a formato Parquet
# df_games_rec.to_parquet('Datasets/Datasets merge/df_games_rec.parquet', index=False)

## df_games_rec_top_3

Habiendo verificado la correctar carga y ordenamiento de los datos en 'df_usr_rec' se filtran y guardan los datos del top 3 de juegos más recomendados por año en el archivo 'df_usr_rec_top_3.csv'.

pd.set_option('display.max_rows', None)

# Mostrar solo los primeros 3 registros por posted_year y descartar los que tienen total_recommend igual a 0
df_games_rec_top_3 = df_games_rec[df_games_rec['total_recommend'] > 0].groupby('posted_year').head(3).reset_index()

# Mostrar df_games_rec_top_100
df_games_rec_top_3

# Exportar a formato XLSX
# df_games_rec_top_3.to_excel('Datasets/Datasets final/df_games_rec_top_3.xlsx', index=False)

# Exportar a formato CSV
df_games_rec_top_3.to_csv('Datasets/Datasets final/df_games_rec_top_3.csv', index=False)

# Exportar a formato Parquet
# df_games_rec_top_3.to_parquet('Datasets/Datasets final/df_games_rec_top_3.parquet', index=False)

## df_dev_rec_worst_3

En este punto se trabajará sobre la información necesaria para responder a la consulta de la función 'def UsersWorstDeveloper' que debe devolver el top 3 de desarrolladoras con juegos menos recomendados para el año dado, calculado a partir de 'recommend' con valor 'False' y 'sentiment' negativos (valor 0).


# Crear una lista de desarrolladores únicos
developers = df_games['developer'].unique()

# Inicializar un diccionario para almacenar los datos
data_dict = {'developer': [], 'posted_year': [], 'neg_rec': []}

# Iterar sobre los desarrolladores
for dev in developers:
    # Filtrar df_games por el desarrollador actual
    dev_games = df_games[df_games['developer'] == dev]

    # Obtener los años únicos de df_reviews para ese desarrollador
    years = df_reviews[df_reviews['game_id'].isin(dev_games['game_id'])]['posted_year'].unique()

    # Iterar sobre los años
    for year in years:
        # Filtrar df_reviews por el desarrollador y el año actual
        dev_year_reviews = df_reviews[(df_reviews['game_id'].isin(dev_games['game_id'])) & (df_reviews['posted_year'] == year)]

        # Calcular la sumatoria de registros donde recommend es False y sentiment es 0
        neg_rec_count = dev_year_reviews[(dev_year_reviews['recommend'] == False) & (dev_year_reviews['sentiment'] == 0)].shape[0]

        # Agregar datos al diccionario
        data_dict['developer'].append(dev)
        data_dict['posted_year'].append(year)
        data_dict['neg_rec'].append(neg_rec_count)

# Crear el DataFrame df_dev_rec
df_dev_rec = pd.DataFrame(data_dict)

# Eliminar filas con developer igual a 'Sin Datos' o posted_year igual a 0
df_dev_rec = df_dev_rec[(df_dev_rec['developer'] != 'Sin Datos') & (df_dev_rec['posted_year'] != 0)]

# Ordenar el DataFrame por neg_rec de forma descendente
df_dev_rec = df_dev_rec.sort_values(by=['posted_year', 'neg_rec'], ascending=[True, False]).reset_index()

# Imprimir el DataFrame resultante
pd.reset_option('display.max_rows')
df_dev_rec

# Exportar a formato XLSX
# df_dev_rec.to_excel('Datasets/Datasets merge/df_dev_rec.xlsx', index=False)

# Exportar a formato CSV
# df_dev_rec.to_csv('Datasets/Datasets merge/df_dev_rec.csv', index=False)

# Exportar a formato Parquet
# df_dev_rec.to_parquet('Datasets/Datasets merge/df_dev_rec.parquet', index=False)

Se genera el datafram con el top 3 de desarrolladores menos recomendados y se guarda en 'df_dev_rec_worst_3.csv'.

# Guardar solo los primeros 3 registros por posted_year y descartar los que tienen neg_rec igual a 0
df_dev_rec_worst_3 = df_dev_rec[df_dev_rec['neg_rec'] > 0].groupby('posted_year').head(3)

# Mostrar df_dev_rec_worst_3
df_dev_rec_worst_3

# Exportar a formato XLSX
# df_dev_rec_worst_3.to_excel('Datasets/Datasets final/df_dev_rec_worst_3.xlsx', index=False)

# Exportar a formato CSV
df_dev_rec_worst_3.to_csv('Datasets/Datasets final/df_dev_rec_worst_3.csv', index=False)

# Exportar a formato Parquet
# df_dev_rec_worst_3.to_parquet('Datasets/Datasets final/df_dev_rec_worst_3.parquet', index=False)

def UsersWorstDeveloper(posted_year: int):
    # Filtrar df_dev_rec_worst_3 por el posted_year proporcionado
    year_data = df_dev_rec_worst_3[df_dev_rec_worst_3['posted_year'] == posted_year]

    if not year_data.empty:
        # Presentar mensaje
        print(f"Desarrolladores con más reseñas negativas para el año {posted_year} son:")

        # Mostrar developer y neg_rec para el posted_year dado
        result = year_data[['developer', 'neg_rec']]
        print(result)
    else:
        print(f"No hay datos disponibles para el posted_year {posted_year} en df_dev_rec_worst_3.")

UsersWorstDeveloper(2015)

## df_dev_sent

Se trabaja sobre la información necesaria para responder a la consulta de la función 'def sentiment_analysis'. Según la empresa desarrolladora, se debe devolver un diccionario con el nombre de la desarrolladora como llave y una lista con la cantidad total de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento como valor. 

Para ello se genera el archivo 'df_dev_sent.csv', contabilizando los valores de comentarios analizados con un valor de sentimiento.


# Realizar la fusión (merge) entre df_reviews y df_games en base al campo game_id
merged_df = pd.merge(df_reviews, df_games, on='game_id', how='left')

# Inicializar un diccionario para almacenar los datos
developer_sentiments = {'developer': [], 'negative_sentiment': [], 'neutral_sentiment': [], 'positive_sentiment': [], 'total_sentiment': []}

# Iterar sobre los desarrolladores únicos
for developer in merged_df['developer'].unique():
    # Filtrar el DataFrame fusionado por el desarrollador actual
    dev_reviews = merged_df[merged_df['developer'] == developer]

    # Calcular la sumatoria de registros para cada categoría de sentimiento
    negative_count = dev_reviews[dev_reviews['sentiment'] == 0].shape[0]
    neutral_count = dev_reviews[dev_reviews['sentiment'] == 1].shape[0]
    positive_count = dev_reviews[dev_reviews['sentiment'] == 2].shape[0]
    total_count = len(dev_reviews)

    # Agregar datos al diccionario
    developer_sentiments['developer'].append(developer)
    developer_sentiments['negative_sentiment'].append(negative_count)
    developer_sentiments['neutral_sentiment'].append(neutral_count)
    developer_sentiments['positive_sentiment'].append(positive_count)
    developer_sentiments['total_sentiment'].append(total_count)

# Crear el DataFrame df_dev_sent
df_dev_sent = pd.DataFrame(developer_sentiments)

# Ordenar el DataFrame por developer alfabéticamente descendente
df_dev_sent = df_dev_sent.sort_values(by='developer', ascending=False)

# Imprimir el DataFrame resultante
df_dev_sent


# Exportar a formato XLSX
# df_dev_sent.to_excel('Datasets/Datasets final/df_dev_sent.xlsx', index=False)

# Exportar a formato CSV
df_dev_sent.to_csv('Datasets/Datasets final/df_dev_sent.csv', index=False)

# Exportar a formato Parquet
# df_dev_sent.to_parquet('Datasets/Datasets final/df_dev_sent.parquet', index=False)

### Exportaciones finales

Se guardan todos los dataframes generados para el deploy de la API en la carpeta correspondiente a la misma.

# Exportar a formato CSV
df_play_gnr_top.to_csv('FastAPI/Datasets/df_play_gnr.csv', index=False)
# Exportar a formato CSV
df_usr_gnr_top.to_csv('FastAPI/Datasets/df_usr_gnr_top.csv', index=False)
# Exportar a formato CSV
df_usr_gnr_top_year.to_csv('FastAPI/Datasets/df_usr_gnr_top_year.csv', index=False)
# Exportar a formato CSV
df_games_rec_top_3.to_csv('FastAPI/Datasets/df_games_rec_top_3.csv', index=False)
# Exportar a formato CSV
df_dev_rec_worst_3.to_csv('FastAPI/Datasets/df_dev_rec_worst_3.csv', index=False)
# Exportar a formato CSV
df_dev_sent.to_csv('FastAPI/Datasets/df_dev_sent.csv', index=False)
