==> 4. Modelo Recomendacion.ipynb <==
#import csv
import pandas as pd
#import numpy as np
import pyarrow as pa
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import StandardScaler

import my_functions
%load_ext autoreload
%autoreload 2

## Carga de datasets a utilizar para generar las funciones

# Cargar el archivo CSV en un DataFrame
df_games = pd.read_csv('Datasets/Datasets ETL/steam_games_v2.csv')
# Cargar el archivo CSV en un DataFrame
df_reviews = pd.read_csv('Datasets/Datasets ETL/user_reviews_v1.csv')
# Cargar el archivo CSV en un DataFrame
df_items = pd.read_csv('Datasets/Datasets ETL/user_items_v2.csv')

## df_games_names

# Seleccionar las columnas 'game_id' y 'app_name' de df_games_data
df_games_names = df_games[['game_id', 'app_name']].copy()

# Convertir el campo 'game_id' a tipo string
df_games_names['game_id'] = df_games_names['game_id'].astype(str)
#df_games_names.set_index('game_id', inplace=True)

df_games_names.info()

# Exportar a formato CSV
df_games_names.to_csv('Datasets/Datasets final/df_games_names.csv', index=False)

# Exportar 
df_games_names.to_csv('FastAPI/Datasets/df_games_names.csv', index=False)

pd.reset_option('display.max_rows')
df_games_names

## Machine Learning

En esta etapa se trabajará sobre los datos necesarios para responder a 2 funciones de recomendación:
- 'def recomendacion_juego': ingresando el id de un producto, deberíamos recibir una lista con 5 juegos recomendados similares al ingresado.

- 'def recomendacion_usuario': ingresando el id de un usuario, deberíamos recibir una lista con 5 juegos recomendados para dicho usuario.

Para ello se aplicará un algoritmo de similitud del coseno, generando matrices de similitud entre juegos y entre usuarios.

## Recomendación de juegos

## df_games_data

# Completar la columna 'app_name'
df_games_data = df_games[['game_id', 'app_name']]

# Completar la columna 'playtime' desde df_items
playtime_by_game = df_items.groupby('game_id')['playtime_forever'].sum()
df_games_data = pd.merge(df_games_data, playtime_by_game, on='game_id', how='left')
df_games_data = df_games_data.rename(columns={'playtime_forever': 'playtime'})

# Completar las columnas 'neg_rec' y 'pos_rec' desde df_reviews
df_reviews['neg_rec'] = ((df_reviews['recommend'] == False) & ((df_reviews['sentiment'] == 0) | df_reviews['sentiment'].isnull())).astype(int)
df_reviews['pos_rec'] = ((df_reviews['recommend'] == True) & ((df_reviews['sentiment'] == 1) | (df_reviews['sentiment'] == 2) | df_reviews['sentiment'].isnull())).astype(int)

# Sumar los valores por 'game_id'
neg_rec_counts = df_reviews.groupby('game_id')['neg_rec'].sum().reset_index()
pos_rec_counts = df_reviews.groupby('game_id')['pos_rec'].sum().reset_index()

# Completar las columnas 'neg_rec' y 'pos_rec' en df_games_data
df_games_data = pd.merge(df_games_data, neg_rec_counts, on='game_id', how='left')
df_games_data = pd.merge(df_games_data, pos_rec_counts, on='game_id', how='left')

# Rellenar con ceros los valores nulos después de la fusión
df_games_data[['neg_rec', 'pos_rec']] = df_games_data[['neg_rec', 'pos_rec']].fillna(0).astype(int)

# Agregar la columna 'total_rec' 
df_games_data['total_rec'] = df_games_data['neg_rec'] + df_games_data['pos_rec']

# Completar las columnas con prefijo 'gnr_' desde df_games
gnr_columns = [col for col in df_games.columns if col.startswith('gnr_')]

# Renombrar las columnas antes de unirlas
# rename_dict = {col: f"{col}_games" for col in gnr_columns}
df_games_data = pd.merge(df_games_data, df_games[['game_id'] + gnr_columns], on='game_id', how='left')

# Completar los campos con valor nulo
# df_games_selection[['app_name']] = df_games_selection[['app_name']].fillna('Sin Datos')
df_games_data = df_games_data.fillna(0)

# Convertir las columnas a los tipos especificados
df_games_data = df_games_data.astype({
    'game_id': 'string',
    'playtime': int
})

# Convertir las columnas gnr_ a tipo int
gnr_columns = [col for col in df_games_data.columns if col.startswith('gnr_')]
df_games_data[gnr_columns] = df_games_data[gnr_columns].astype(int)

# Resetear indice
df_games_data = df_games_data.reset_index(drop=True)
df_games_data.info()

Reservamos información actual de df_games_data en df_games_selection para luego trabajar sobre esté último.

# Obtener las columnas que comienzan con 'gnr_'
gnr_columns = [col for col in df_games_data.columns if col.startswith('gnr_')]

# Filtrar las filas donde todas las columnas 'gnr_' tengan valor 0
df_games_selection = df_games_data[~(df_games_data[gnr_columns] == 0).all(axis=1)]

Continuamos trabajando sobre df_games_data.

# Tomar solo las columnas necesarias de df_games
cols_to_keep = ['game_id', 'app_name', 'playtime'] + [col for col in df_games.columns if col.startswith('gnr_')]
df_games_data = df_games_data[cols_to_keep].copy()

# Filtrar filas donde todas las columnas 'gnr_' sean 0
df_games_data = df_games_data[df_games_data.filter(like='gnr_').sum(axis=1) > 0]

# Establecer 'game_id' como índice
# df_games_data.set_index('game_id', inplace=True)

# Resetear indice
df_games_data = df_games_data.reset_index(drop=True)


# Calcular el percentil 85 de playtime
playtime_threshold = df_games_data['playtime'].quantile(0.85)

# Filtrar las filas con playtime mayor que el percentil 85
df_games_data = df_games_data[df_games_data['playtime'] > playtime_threshold].reset_index(drop=True)


# Eliminar la columna 'playtime'
df_games_data = df_games_data.drop(columns=['playtime'])

df_games_data.info()

## Filtros

## df_games_selection

Primero se trabaja sobre el dataset con la información necesaria para la comparación entre juegos.

df_games_selection.info()

# Eliminar filas donde neg_rec > pos_rec
df_games_selection = df_games_selection[df_games_selection['neg_rec'] < df_games_selection['pos_rec']]

df_games_selection.info()

# Filtrar y mantener solo las filas donde playtime no sea 0
df_games_selection = df_games_selection[df_games_selection['playtime'] > 0]
df_games_selection.info()

# Seleccionar columnas 'app_name' y 'gnr_'
cols_to_select = ['game_id', 'app_name'] + [col for col in df_games_selection.columns if col.startswith('gnr_')]
df_games_selection = df_games_selection[cols_to_select].copy()
df_games_selection = df_games_selection.reset_index(drop=True)

df_games_selection.info()

df_games_data.info()

## Modelo

df_games_data.shape

df_games_selection.shape

## df_games_similarity

cols_for_similarity = [col for col in df_games_data.columns if col.startswith('gnr_')]

# Seleccionar las filas que contienen datos para las columnas seleccionadas
df_games_data_model = df_games_data[['game_id'] + cols_for_similarity]

# Seleccionar las columnas que contienen datos para las columnas seleccionadas
df_games_selection_model = df_games_selection[['game_id'] + cols_for_similarity]

# Establecer 'game_id' como índice
df_games_data_model.set_index('game_id', inplace=True)
df_games_selection_model.set_index('game_id', inplace=True)

# Calcular la matriz de similitud del coseno
similarity_matrix = cosine_similarity(df_games_data_model, df_games_selection_model)

# Crear un DataFrame con la matriz de similitud
df_games_similarity = pd.DataFrame(similarity_matrix, index=df_games_data_model.index, columns=df_games_selection_model.index)

df_games_similarity.info()

df_games_similarity.head(5)

# Exportar a formato Parquet manteniendo los índices
df_games_similarity.to_parquet('Datasets/Datasets final/df_games_similarity.parquet', index=True)

# Exportar a formato Parquet manteniendo los índices
df_games_similarity.to_parquet('FastAPi/Datasets/df_games_similarity.parquet', index=True)

# Cargar el archivo csv en un DataFrame
df_games_names = pd.read_csv('FastAPI/Datasets/df_games_names.csv')

# Convertir el campo 'game_id' a tipo string
df_games_names['game_id'] = df_games_names['game_id'].astype(str)

df_games_names.info()

## def recomendacion_juego

def recomendacion_juego(game_id:str, num_recommendations=5):

    game_name = df_games_names.loc[df_games_names['game_id'] == game_id, 'app_name'].values[0]

    # Convertir el campo 'game_id' a tipo string
    df_games_names['game_id'] = df_games_names['game_id'].astype(str)

    # Obtener la serie de juegos similares, excluir el game_id introducido 
    similar_games = df_games_similarity.drop(game_id).loc[:, game_id].sort_values(ascending=False)

    # Tomar las primeras num_recommendations recomendaciones
    recommendations = similar_games.head(num_recommendations).index.tolist()

    recommended_games = df_games_names.loc[df_games_names['game_id'].isin(recommendations), ['game_id', 'app_name']].reset_index(drop=True)

    recommended_games = recommended_games[['game_id', 'app_name']].to_dict(orient='records')

    message = f"Juegos similares a {game_name} (game_id {game_id}):"

    result = f"{message}", recommended_games
    
    return result


# Ejemplo de uso
game_id_input = '282010'
recommendations = recomendacion_juego(game_id_input)
recommendations

## Recomendación usuario - item

df_items.head(1)

my_functions.describe_df(df_items)

## df_usr_data

# Seleccionar solo las columnas requeridas
df_usr_data = df_items[['game_id', 'playtime_forever', 'user_id']]

# Convertir la columna 'game_id' a tipo de dato str
df_usr_data['game_id'] = df_usr_data['game_id'].astype(str)

# Ordenar por user_id y playtime_forever
df_usr_data = df_usr_data.sort_values(by=['user_id', 'playtime_forever'], ascending=[True, False]).groupby('user_id').head(3)


# Configurar la opción para mostrar todas las filas
pd.set_option('display.max_rows', None)
#pd.reset_option('display.max_rows')

# Imprimir el DataFrame resultante
df_usr_data.head(15)

my_functions.describe_df(df_usr_data)

# Eliminar la columna 'playtime_forever'
df_usr_data = df_usr_data.drop('playtime_forever', axis=1)


my_functions.describe_df(df_usr_data)

# Agrupar por 'user_id' y contar la cantidad de registros para cada grupo
df_usr_data_count = df_usr_data.groupby('user_id').count().reset_index()

pd.reset_option('display.max_rows')
df_usr_data_count

# Borrar registros en df_usr_data_count donde game_id sea menor a 3
df_usr_data_count = df_usr_data_count[df_usr_data_count['game_id'] >= 3]

my_functions.describe_df(df_usr_data_count)

# Quedarse solo con registros en df_usr_data que compartan user_id con df_usr_data_count_filtered
df_usr_data = df_usr_data[df_usr_data['user_id'].isin(df_usr_data_count['user_id'])]

my_functions.describe_df(df_usr_data)

df_usr_data

# Quedarse con los primeros 3000 usuarios
df_usr_data = df_usr_data.head(9000)

df_usr_data

my_functions.describe_df(df_usr_data)

pd.reset_option('display.max_rows')

# Exportar 
df_usr_data.to_parquet('Datasets/Datasets final/df_usr_data.parquet', index=False)

# Exportar 
df_usr_data.to_parquet('FastAPI/Datasets/df_usr_data.parquet', index=False)

## df_usr_selection

# Quedarse con el primer registro de cada user_id
df_usr_selection = df_usr_data.drop_duplicates(subset='user_id', keep='first')

# Imprimir el DataFrame resultante
df_usr_selection.head(15)

df_usr_selection.info()

# Realizar la concatenación y el filtrado
df_usr_selection = pd.merge(df_usr_selection, df_games_selection[['game_id']], on='game_id', how='inner')


df_usr_selection.info()

df_usr_selection

my_functions.describe_df(df_usr_selection)

my_functions.describe_df(df_usr_data)

## Modelo

# Agrupar por 'user_id' y generar una lista de distintos valores de 'game_id' para cada usuario
df_usr_data_model = df_usr_data.groupby('user_id')['game_id'].agg(lambda x: list(set(x))).reset_index()

# Renombrar la columna
df_usr_data_model.rename(columns={'game_id': 'games'}, inplace=True)


df_usr_data_model

# Establecer 'user_id' como índice
df_usr_data_model.set_index('user_id', inplace=True)

df_usr_data_model

import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import MultiLabelBinarizer

# Convertir la columna 'games' a un formato adecuado para la similitud del coseno
mlb = MultiLabelBinarizer()
usr_matrix = mlb.fit_transform(df_usr_data_model['games'])
usr_df = pd.DataFrame(usr_matrix, index=df_usr_data_model.index)


# Calcular la similitud del coseno entre usuarios
cosine_sim = cosine_similarity(usr_df)

# Crear un DataFrame de similitud del coseno
df_user_similarity = pd.DataFrame(cosine_sim, index=df_usr_data_model.index, columns=df_usr_data_model.index)

pd.set_option('display.max_rows', None)

# Imprimir el DataFrame de similitud del coseno
df_user_similarity.head(10)

# Exportar 
df_user_similarity.to_parquet('Datasets/Datasets final/df_user_similarity.parquet', index=True)

# Exportar 
df_user_similarity.to_parquet('FastAPI/Datasets/df_user_similarity.parquet', index=True)

df_user_similarity.shape

## def recomendacion_usuario

pd.reset_option('display.max_rows')

# Cargar el archivo Parquet en un DataFrame
df_usr_data = pd.read_parquet('FastAPI/Datasets/df_usr_data.parquet')

def recomendacion_usuario(user_id: str, num_recommendations = 5):
    # Filtrar df_usr_data por el user_id específico
    user_games = df_usr_data[df_usr_data['user_id'] == user_id]

    # Crear la lista user_games a partir de los valores de df_usr_data
    user_games = user_games['game_id'].tolist()

    # Obtener la serie de usuarios similares, excluir el game_id introducido 
    similar_users = df_user_similarity.drop(user_id).loc[:, user_id].sort_values(ascending=False).index.tolist()

    # Crear DataFrame con similar_users
    df_similar_users = pd.DataFrame({'user_id': similar_users})

    # Realizar la concatenación con df_usr_selection usando el método merge
    df_concatenado = pd.merge(df_similar_users, df_usr_data, on='user_id', how='inner')

    # Filtrar las filas donde el game_id no está en user_games
    df_concatenado = df_concatenado[~df_concatenado['game_id'].isin(user_games)]

    # Utilizar merge para concatenar basándote en la columna game_id
    df_final = pd.merge(df_concatenado, df_games_names, on='game_id', how='inner')

    # Eliminar duplicados y quedarse con el primer registro
    df_final_sin_duplicados = df_final.drop_duplicates(subset='game_id', keep='first').head(num_recommendations) 
        
    recommended_games = df_final_sin_duplicados[['game_id', 'app_name', 'user_id']].to_dict(orient='records')
 
    message = f"A usuarios similares a {user_id} les gustó:"
    
    #result = message, recommended_games
    result = f"{message}", recommended_games
    
    return result

user_id = '-Encore-'
rec_usuario = recomendacion_usuario(user_id)
rec_usuario
